+++ app/models/dorsale/application_record.rb
@@ -1,8 +1,14 @@
+require "dorsale/search"
+
 class Dorsale::ApplicationRecord < ActiveRecord::Base
   self.abstract_class = true
 
   include Dorsale::ActiveRecordUUIDConcern
   include Dorsale::DefaultValuesConcern
+  include Dorsale::ModelToS
+  include Dorsale::ModelI18n
+  include Dorsale::PolymorphicId
+
   nilify_blanks
 
   def self.last_created